backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
bcrypt
python-jose
pytest
python-multipart
jinja2
reportlab
